requirements.txt
Flask==2.2.2
Flask-Cors==2.1.3
openai==0.28
requests==2.28.2
python-dotenv==0.21.0
